requirements.txt
# core requirements
bottle==0.12.9
bottle-sqlalchemy==0.4.3
requests-oauthlib==0.6.2
WTForms==2.1
git+https://github.com/arahayrabedian/wtforms-nocaptcha#egg=wtforms-nocaptcha
git+https://github.com/arahayrabedian/gmail-sender#egg=gmail-sender
pytz==2017.3
pyOpenSSL==17.5.0

# where your own discretion is required:
psycopg2==2.6.2  # your mileage may vary, need this for implementing pg backing store
